refactor: remove usage of isNull from lodash
