--- src/store/features/transactions/saga/index.ts
@@ -1,1 +1,0 @@
-export { default } from './saga';
