--- src/store/features/transactions/saga/transactions/index.ts
@@ -1,1 +1,0 @@
-export { default } from './transactionsSaga';
