+++ src/store/features/transactions/saga/transactions/transactionsSaga.ts
@@ -1,14 +1,14 @@
 import { takeLatest } from 'redux-saga/effects';
 
-import burnSaga from './burnSaga';
-import mintSaga from './mintSaga';
-import rolloverMintSaga from './rolloverMintSaga';
-import rolloverSwapSaga from './rolloverSwapSaga';
-import settlePositionSaga from './settlePositionSaga';
-import swapSaga from './swapSaga';
-import updatePositionMarginSaga from './updatePositionMarginSaga';
+import { burnSaga } from './burnSaga';
+import { mintSaga } from './mintSaga';
+import { rolloverMintSaga } from './rolloverMintSaga';
+import { rolloverSwapSaga } from './rolloverSwapSaga';
+import { settlePositionSaga } from './settlePositionSaga';
+import { swapSaga } from './swapSaga';
+import { updatePositionMarginSaga } from './updatePositionMarginSaga';
 
-function* transactionsSaga() {
+export function* transactionsSaga() {
   yield takeLatest('mint', mintSaga);
   yield takeLatest('swap', swapSaga);
   yield takeLatest('updatePositionMargin', updatePositionMarginSaga);
@@ -17,5 +17,3 @@ function* transactionsSaga() {
   yield takeLatest('rolloverMint', rolloverMintSaga);
   yield takeLatest('rolloverSwap', rolloverSwapSaga);
 }
-
-export default transactionsSaga;
