+++ src/store/features/transactions/utilities/index.ts
@@ -1,4 +1,4 @@
 export { default as createId } from './createId';
 export { default as deserializeAmm } from './deserializeAmm';
-export { default as getSigner } from './getSigner';
+export * from './getSigner';
 export { default as serializeAmm } from './serializeAmm';
